Update Architectural-design.md
In the 2009 Woolley study, a group member was removed at the 20 minute mark. However, I did not add this since it was not the study listed in the spreadsheet, even though it is were most of the procedure came from.
I've emailed the author for the video shown to participants.